apps/agents/requirements.txt
requests==2.32.5
pillow==11.3.0
serpapi==0.1.5
streamlit==1.50.0
boto3==1.35.64
ultralytics==8.3.50
